Feature/L01-75 (#14)
* feature L01-75 changed PUBLISHED to STATUS enum in Product class

* chore L01-75 removed unused field categoryMapper

* feature L01-75 added activateProductByUuid controller

* chore L01-75 removed 'public' modifier

* feature L01-75 changed updateProductByUuid

* feature L01-75 applied abstract factory pattern

* feature L01-75 added archiveAt field for product

* feature L01-75 added /product/archive endpoint, restored update feature

* feature L01-75: tests added

* chore L01-75: never used variable removed

* feature L01-75: changed Abstract Factory ProductArchiver into ProductArchiverService

* chore L01-90: removed unused variable

* feature L01-75: modified deleteProductByUuId: IN_PREPARATION status delete, ACTIVE status archive

+++ src/main/java/com/liftoff/project/service/impl/ProductServiceImpl.java
@@ -8,6 +8,7 @@ import com.liftoff.project.exception.ProductNotFoundException;
 import com.liftoff.project.mapper.ProductMapper;
 import com.liftoff.project.model.Category;
 import com.liftoff.project.model.Product;
+import com.liftoff.project.model.ProductStatus;
 import com.liftoff.project.repository.CategoryRepository;
 import com.liftoff.project.repository.ProductRepository;
 import com.liftoff.project.service.CategoryService;
@@ -18,8 +19,8 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.time.Instant;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -32,6 +33,7 @@ public class ProductServiceImpl implements ProductService {
     private final CategoryRepository categoryRepository;
     private final ProductMapper productMapper;
     private final CategoryService categoryService;
+    private final ProductArchiverService productArchiverService;
 
     @Override
     public PaginatedProductResponseDTO getProducts(int page, int size) {
@@ -89,42 +91,39 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public ProductResponseDTO updateProductByUuid(UUID productUuid, ProductRequestDTO productRequestDTO) {
-        Optional<Product> productOptional = productRepository.findByUId(productUuid);
-        if (productOptional.isPresent()) {
-            Product existingProduct = productOptional.get();
-            updateProductFromRequest(existingProduct, productRequestDTO);
+        Product existingProduct = productRepository.findByUId(productUuid)
+                .orElseThrow(() ->
+                        new ProductNotFoundException("Product with UUID " + productUuid + " not found."));
 
-            Set<Category> existingCategories = categoryService.getExistingCategories(productRequestDTO.getCategories());
-            existingProduct.setCategories(existingCategories);
+        Set<Category> existingCategories =
+                categoryService.getExistingCategories(productRequestDTO.getCategories());
 
-            Product updatedProduct = productRepository.save(existingProduct);
-            return productMapper.mapEntityToResponse(updatedProduct);
-        } else {
-            String errorMessage = "Product with UUID " + productUuid + " not found.";
-            throw new ProductNotFoundException(errorMessage);
-        }
+        return productArchiverService.archiveProduct(existingProduct, productRequestDTO, existingCategories);
     }
 
     @Override
     public void deleteProductByUuId(UUID productUuId) {
         Product product = productRepository.findByUId(productUuId)
                 .orElseThrow(() -> new ProductNotFoundException("Product with UUID " + productUuId + " not found."));
-        productRepository.delete(product);
+
+        if (product.getStatus() == ProductStatus.IN_PREPARATION) {
+            productRepository.delete(product);
+        } else if (product.getStatus() == ProductStatus.ACTIVE) {
+            product.setStatus(ProductStatus.CLOSED);
+            product.setArchivedAt(Instant.now());
+            productRepository.save(product);
+        }
+
     }
 
-    private void updateProductFromRequest(Product product, ProductRequestDTO productRequestDTO) {
-        product.setName(productRequestDTO.getName());
-        product.setSku(productRequestDTO.getSku());
-        product.setDescription(productRequestDTO.getDescription());
-        product.setRegularPrice(productRequestDTO.getRegularPrice());
-        product.setDiscountPrice(productRequestDTO.getDiscountPrice());
-        product.setDiscountPriceEndDate(productRequestDTO.getDiscountPriceEndDate());
-        product.setLowestPrice(productRequestDTO.getLowestPrice());
-        product.setShortDescription(productRequestDTO.getShortDescription());
-        product.setNote(productRequestDTO.getNote());
-        product.setPublished(productRequestDTO.getPublished());
-        product.setProductscol(productRequestDTO.getProductscol());
-        product.setQuantity(productRequestDTO.getQuantity());
+    @Override
+    public ProductResponseDTO activateProduct(ProductResponseDTO product) {
+        if (product.getStatus().equals(ProductStatus.ACTIVE)) return product;
+
+        product.setStatus(ProductStatus.ACTIVE);
+        product.setUpdatedAt(Instant.now());
+
+        return product;
     }
 
 }
